+++ src/FSharp.Literate/Main.fs
@@ -1,5 +1,6 @@
 ﻿namespace FSharp.Literate
 
+open System
 open System.IO
 open System.Reflection
 open FSharp.Markdown
@@ -97,17 +98,22 @@ type Literate private () =
   static member WriteHtml(doc:LiterateDocument, ?prefix, ?lineNumbers) =
     let ctx = formattingContext None (Some OutputKind.Html) prefix lineNumbers None None None
     let doc = Transformations.replaceLiterateParagraphs ctx doc
-    Markdown.WriteHtml(MarkdownDocument(doc.Paragraphs @ [InlineBlock doc.FormattedTips], doc.DefinedLinks))
+    let doc = MarkdownDocument(doc.Paragraphs @ [InlineBlock doc.FormattedTips], doc.DefinedLinks)
+    let sb = new System.Text.StringBuilder()
+    use wr = new StringWriter(sb)
+    Html.formatMarkdown wr Environment.NewLine true doc.DefinedLinks doc.Paragraphs
+    sb.ToString()
+
+  static member WriteHtml(doc:LiterateDocument, writer:TextWriter, ?prefix, ?lineNumbers) =
+    let ctx = formattingContext None (Some OutputKind.Html) prefix lineNumbers None None None
+    let doc = Transformations.replaceLiterateParagraphs ctx doc
+    let doc = MarkdownDocument(doc.Paragraphs @ [InlineBlock doc.FormattedTips], doc.DefinedLinks)
+    Html.formatMarkdown writer Environment.NewLine true doc.DefinedLinks doc.Paragraphs
 
   static member WriteLatex(doc:LiterateDocument, ?prefix, ?lineNumbers) =
     let ctx = formattingContext None (Some OutputKind.Latex) prefix lineNumbers None None None
     let doc = Transformations.replaceLiterateParagraphs ctx doc
     Markdown.WriteLatex(MarkdownDocument(doc.Paragraphs, doc.DefinedLinks))
-
-  static member WriteHtml(doc:LiterateDocument, writer:TextWriter, ?prefix, ?lineNumbers) =
-    let ctx = formattingContext None (Some OutputKind.Html) prefix lineNumbers None None None
-    let doc = Transformations.replaceLiterateParagraphs ctx doc
-    Markdown.WriteHtml(MarkdownDocument(doc.Paragraphs @ [InlineBlock doc.FormattedTips], doc.DefinedLinks), writer)
 
   static member WriteLatex(doc:LiterateDocument, writer:TextWriter, ?prefix, ?lineNumbers) =
     let ctx = formattingContext None (Some OutputKind.Latex) prefix lineNumbers None None None