+++ RELEASE_NOTES.md
@@ -49,3 +49,7 @@
 * 2.4.11 - Improved handling of end comments and http/https links at end of lines
 * 2.4.12 - Be more flexible about URL generating, protect against exceptions, fix bugs and typos
 * 2.4.13 - Fix evaluation bug; Return characters for horizontal lines
+* 2.4.14 - Update NuGet dependencies
+* 2.4.15 - Support combination of commands (e.g. hide, define-output), add do-not-eval-file
+* 2.4.16 - Improve formatting of literate scripts (generate tables around pre)
+* 2.4.17 - Fix comment parsing when whitespace